#4920 Force Initial IFA import process to be a java process
#4920 Improve creating product prices when importing IFA products
